SFLE/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0  # для работы с изображениями